+++ graph/src/components/ethereum/types.rs
@@ -1,9 +1,7 @@
-use ethabi::LogParam;
 use serde::{Deserialize, Serialize};
 use std::{convert::TryFrom, sync::Arc};
 use web3::types::{
-    Action, Address, Block, Bytes, Log, Res, Trace, Transaction, TransactionReceipt, H160, H256,
-    U128, U256, U64,
+    Action, Address, Block, Bytes, Log, Res, Trace, Transaction, TransactionReceipt, H256, U256,
 };
 
 use crate::{
@@ -167,145 +165,6 @@ impl EthereumCall {
             transaction_hash: trace.transaction_hash,
             transaction_index,
         })
-    }
-}
-
-/// Ethereum block data.
-#[derive(Clone, Debug, Default)]
-pub struct EthereumBlockData {
-    pub hash: H256,
-    pub parent_hash: H256,
-    pub uncles_hash: H256,
-    pub author: H160,
-    pub state_root: H256,
-    pub transactions_root: H256,
-    pub receipts_root: H256,
-    pub number: U64,
-    pub gas_used: U256,
-    pub gas_limit: U256,
-    pub timestamp: U256,
-    pub difficulty: U256,
-    pub total_difficulty: U256,
-    pub size: Option<U256>,
-}
-
-impl<'a, T> From<&'a Block<T>> for EthereumBlockData {
-    fn from(block: &'a Block<T>) -> EthereumBlockData {
-        EthereumBlockData {
-            hash: block.hash.unwrap(),
-            parent_hash: block.parent_hash,
-            uncles_hash: block.uncles_hash,
-            author: block.author,
-            state_root: block.state_root,
-            transactions_root: block.transactions_root,
-            receipts_root: block.receipts_root,
-            number: block.number.unwrap(),
-            gas_used: block.gas_used,
-            gas_limit: block.gas_limit,
-            timestamp: block.timestamp,
-            difficulty: block.difficulty,
-            total_difficulty: block.total_difficulty.unwrap_or_default(),
-            size: block.size,
-        }
-    }
-}
-
-/// Ethereum transaction data.
-#[derive(Clone, Debug)]
-pub struct EthereumTransactionData {
-    pub hash: H256,
-    pub index: U128,
-    pub from: H160,
-    pub to: Option<H160>,
-    pub value: U256,
-    pub gas_used: U256,
-    pub gas_price: U256,
-    pub input: Bytes,
-}
-
-impl From<&'_ Transaction> for EthereumTransactionData {
-    fn from(tx: &Transaction) -> EthereumTransactionData {
-        EthereumTransactionData {
-            hash: tx.hash,
-            index: tx.transaction_index.unwrap().as_u64().into(),
-            from: tx.from,
-            to: tx.to,
-            value: tx.value,
-            gas_used: tx.gas,
-            gas_price: tx.gas_price,
-            input: tx.input.clone(),
-        }
-    }
-}
-
-/// An Ethereum event logged from a specific contract address and block.
-#[derive(Debug)]
-pub struct EthereumEventData {
-    pub address: Address,
-    pub log_index: U256,
-    pub transaction_log_index: U256,
-    pub log_type: Option<String>,
-    pub block: EthereumBlockData,
-    pub transaction: EthereumTransactionData,
-    pub params: Vec<LogParam>,
-}
-
-impl Clone for EthereumEventData {
-    fn clone(&self) -> Self {
-        EthereumEventData {
-            address: self.address,
-            log_index: self.log_index,
-            transaction_log_index: self.transaction_log_index,
-            log_type: self.log_type.clone(),
-            block: self.block.clone(),
-            transaction: self.transaction.clone(),
-            params: self
-                .params
-                .iter()
-                .map(|log_param| LogParam {
-                    name: log_param.name.clone(),
-                    value: log_param.value.clone(),
-                })
-                .collect(),
-        }
-    }
-}
-
-/// An Ethereum call executed within a transaction within a block to a contract address.
-#[derive(Debug)]
-pub struct EthereumCallData {
-    pub from: Address,
-    pub to: Address,
-    pub block: EthereumBlockData,
-    pub transaction: EthereumTransactionData,
-    pub inputs: Vec<LogParam>,
-    pub outputs: Vec<LogParam>,
-}
-
-impl Clone for EthereumCallData {
-    fn clone(&self) -> Self {
-        EthereumCallData {
-            to: self.to,
-            from: self.from,
-            block: self.block.clone(),
-            transaction: self.transaction.clone(),
-            inputs: self
-                .inputs
-                .iter()
-                .map(|log_param| LogParam {
-                    name: log_param.name.clone(),
-                    value: log_param.value.clone(),
-                })
-                .collect(),
-            outputs: self
-                .outputs
-                .iter()
-                .map(|log_param| LogParam {
-                    name: log_param.name.clone(),
-                    value: log_param.value.clone(),
-                })
-                .collect(),
-        }
     }
 }
