+++ graph/src/data/subgraph/mod.rs
@@ -593,13 +593,6 @@ impl Mapping {
             .any(|handler| matches!(handler.filter, Some(BlockHandlerFilter::Call)))
     }
 
-    pub fn required_capabilities(&self) -> NodeCapabilities {
-        NodeCapabilities {
-            traces: self.has_block_handler_with_call_filter() || self.has_call_handler(),
-            archive: self.calls_host_fn("ethereum.call"),
-        }
-    }
-
     pub fn find_abi(&self, abi_name: &str) -> Result<Arc<MappingABI>, Error> {
         Ok(self
             .abis
