+++ runtime/derive/Cargo.toml
@@ -9,3 +9,4 @@ proc-macro = true
 [dependencies]
 syn = { version = "1.0", features = ["full"] }
 quote = "1.0"
+anyhow = "1.0"
