+++ runtime/wasm/src/lib.rs
@@ -1,4 +1,4 @@
-mod asc_abi;
+pub mod asc_abi;
 mod to_from;
 
 /// Public interface of the crate, receives triggers to be processed.
@@ -15,18 +15,3 @@ mod module;
 mod host_exports;
 
 mod error;
-
-use graph::prelude::web3::types::Address;
-use graph::prelude::SubgraphStore;
-
-#[derive(Clone, Debug)]
-pub(crate) struct UnresolvedContractCall {
-    pub contract_name: String,
-    pub contract_address: Address,
-    pub function_name: String,
-    pub function_signature: Option<String>,
-    pub function_args: Vec<ethabi::Token>,
-}
-
-trait RuntimeStore: SubgraphStore {}
-impl<S: SubgraphStore> RuntimeStore for S {}
